refactor: move logging processors from v4 to "legacy"
