+++ gravitee-apim-gateway/gravitee-apim-gateway-handlers/gravitee-apim-gateway-handlers-api/src/test/java/io/gravitee/gateway/reactive/handlers/api/processor/AbstractProcessorTest.java
@@ -44,6 +44,9 @@ public class AbstractProcessorTest {
     @Mock
     protected MutableRequest mockRequest;
 
+    @Mock
+    protected Metrics mockMetrics;
+
     protected HttpHeaders spyRequestHeaders;
 
     @Spy
